automation-framework/tests/test_regression.py
"""
Comprehensive regression tests for automation-framework module.
These tests verify core functionality and generate detailed Allure reports.
"""

import pytest
import allure
import json
import tempfile
from pathlib import Path
from unittest.mock import patch, MagicMock

# Import the actual modules being tested
from src.api.client import APIClient
from src.config.settings import FrameworkConfig
from src.core.base_page import BasePage
from src.data.factories import UserFactory, ProductFactory, OrderFactory, DataManager
from src.utils.helpers import generate_random_string, format_timestamp, validate_email
from src.utils.logger import ContextLogger, get_logger


@allure.epic("Automation Framework")
@allure.feature("Regression Testing")
class TestAutomationFrameworkRegression:
    """Comprehensive regression tests for automation framework components."""

    @allure.story("API Client")
    @allure.title("Test API Client initialization and configuration")
    @allure.description("Verify API client can be initialized with proper configuration.")
    @allure.severity(allure.severity_level.CRITICAL)
    @allure.tag("api", "client", "regression")
    def test_api_client_initialization(self):
        """Test API client initialization."""
        with allure.step("Initialize API client"):
            client = APIClient()
            assert client is not None
            allure.attach("API client initialized successfully", "Status", allure.attachment_type.TEXT)

    @allure.story("API Client")
    @allure.title("Test API client configuration methods")
    @allure.description("Verify API client configuration methods work correctly.")
    @allure.severity(allure.severity_level.NORMAL)
    @allure.tag("api", "config", "regression")
    def test_api_client_configuration(self):
        """Test API client configuration."""
        with allure.step("Create API client with custom config"):
            client = APIClient()
            
        with allure.step("Test configuration methods"):
            # Test basic configuration
            assert hasattr(client, 'base_url')
            assert hasattr(client, 'timeout')
            
            allure.attach("Configuration methods verified", "Status", allure.attachment_type.TEXT)

    @allure.story("Settings")
    @allure.title("Test settings configuration loading")
    @allure.description("Verify settings can be loaded and configured properly.")
    @allure.severity(allure.severity_level.CRITICAL)
    @allure.tag("config", "settings", "regression")
    def test_settings_loading(self):
        """Test settings configuration."""
        with allure.step("Load settings"):
            settings = FrameworkConfig()
            assert settings is not None
            
        with allure.step("Verify settings properties"):
            assert hasattr(settings, 'test')
            assert hasattr(settings, 'browser')
            assert hasattr(settings, 'api')
            assert hasattr(settings, 'database')
            
            allure.attach(f"Settings loaded: {settings.__dict__}", "Settings", allure.attachment_type.JSON)

    @allure.story("Base Page")
    @allure.title("Test base page functionality")
    @allure.description("Verify base page class provides expected functionality.")
    @allure.severity(allure.severity_level.NORMAL)
    @allure.tag("page", "base", "regression")
    def test_base_page_functionality(self):
        """Test base page functionality."""
        with allure.step("Initialize base page with mock driver"):
            # Mock driver for testing
            mock_driver = MagicMock()
            page = BasePage(mock_driver)
            assert page is not None
            
        with allure.step("Test base page methods"):
            assert hasattr(page, 'wait_for_element_visible')
            assert hasattr(page, 'click_element')
            assert hasattr(page, 'get_text')
            assert hasattr(page, 'find_element')
            assert hasattr(page, 'navigate_to')
            
            allure.attach("Base page methods verified", "Status", allure.attachment_type.TEXT)

    @allure.story("Data Factory")
    @allure.title("Test data factory generation")
    @allure.description("Verify data factory can generate test data correctly.")
    @allure.severity(allure.severity_level.NORMAL)
    @allure.tag("data", "factory", "regression")
    def test_data_factory_generation(self):
        """Test data factory functionality."""
        with allure.step("Initialize data factory"):
            factory = DataManager()
            assert factory is not None
            
        with allure.step("Test data generation"):
            # Test generating sample data
            sample_data = factory.generate_test_dataset(10, 5, 10)
            assert sample_data is not None
            assert isinstance(sample_data, dict)
            assert "users" in sample_data
            assert "products" in sample_data
            assert "orders" in sample_data
            
            allure.attach(json.dumps(sample_data, indent=2), "Generated Data", allure.attachment_type.JSON)

    @allure.story("Helper Utilities")
    @allure.title("Test helper utility functions")
    @allure.description("Verify helper utilities work as expected.")
    @allure.severity(allure.severity_level.NORMAL)
    @allure.tag("utils", "helpers", "regression")
    def test_helper_utilities(self):
        """Test helper utility functions."""
        with allure.step("Test helper functions"):
            # Test generate_random_string function
            random_string = generate_random_string(10)
            assert len(random_string) == 10
            assert isinstance(random_string, str)
            
            # Test format_timestamp function
            timestamp = format_timestamp()
            assert isinstance(timestamp, str)
            assert len(timestamp) > 0
            
            # Test validate_email function
            valid_email = validate_email("test@example.com")
            assert valid_email is True
            
            invalid_email = validate_email("invalid-email")
            assert invalid_email is False
            
            allure.attach(f"Generated random string: {random_string}", "Helper Test", allure.attachment_type.TEXT)

    @allure.story("Logger")
    @allure.title("Test logging functionality")
    @allure.description("Verify logging system works correctly.")
    @allure.severity(allure.severity_level.NORMAL)
    @allure.tag("logging", "regression")
    def test_logging_functionality(self):
        """Test logging functionality."""
        with allure.step("Initialize logger"):
            logger = get_logger("test_regression")
            assert logger is not None
            
        with allure.step("Test logging methods"):
            assert hasattr(logger, 'info')
            assert hasattr(logger, 'error')
            assert hasattr(logger, 'debug')
            
            # Test logging (should not raise exceptions)
            logger.info("Test log message")
            logger.error("Test error message")
            logger.debug("Test debug message")
            
            allure.attach("Logging methods tested successfully", "Status", allure.attachment_type.TEXT)

    @allure.story("Integration")
    @allure.title("Test module integration")
    @allure.description("Verify all modules work together correctly.")
    @allure.severity(allure.severity_level.CRITICAL)
    @allure.tag("integration", "regression")
    def test_module_integration(self):
        """Test integration between modules."""
        with allure.step("Initialize all components"):
            settings = FrameworkConfig()
            client = APIClient()
            mock_driver = MagicMock()
            page = BasePage(mock_driver)
            factory = DataManager()
            logger = get_logger("integration_test")
            
        with allure.step("Verify component interaction"):
            # Test that components can work together
            test_data = factory.generate_test_dataset(5, 3, 5)
            logger.info(f"Generated test data: {test_data}")
            
            # Test helper functions
            random_string = generate_random_string(10)
            timestamp = format_timestamp()
            
            # Verify all components are properly initialized
            assert settings is not None
            assert client is not None
            assert page is not None
            assert factory is not None
            assert logger is not None
            assert random_string is not None
            assert timestamp is not None
            
            allure.attach("All components integrated successfully", "Integration", allure.attachment_type.TEXT)

    @allure.story("Error Handling")
    @allure.title("Test error handling and recovery")
    @allure.description("Verify proper error handling across the framework.")
    @allure.severity(allure.severity_level.NORMAL)
    @allure.tag("error-handling", "regression")
    def test_error_handling(self):
        """Test error handling functionality."""
        with allure.step("Test error scenarios"):
            # Test with invalid inputs using helper functions
            try:
                # This should handle errors gracefully
                result = generate_random_string(-1)
                # If it doesn't raise an exception, it should return something reasonable
                assert result is not None
            except Exception as e:
                # If it raises an exception, it should be a reasonable one
                assert isinstance(e, (ValueError, AssertionError))
                
            # Test email validation with invalid input
            try:
                invalid_email = validate_email("invalid-email")
                assert invalid_email is False
            except Exception as e:
                # Should handle gracefully
                assert isinstance(e, (ValueError, TypeError))
                
            allure.attach("Error handling tested successfully", "Status", allure.attachment_type.TEXT)

    @allure.story("Performance")
    @allure.title("Test performance characteristics")
    @allure.description("Verify framework performance meets expectations.")
    @allure.severity(allure.severity_level.NORMAL)
    @allure.tag("performance", "regression")
    def test_performance_characteristics(self):
        """Test performance characteristics."""
        import time
        
        with allure.step("Measure initialization time"):
            start_time = time.time()
            
            # Initialize all components
            settings = FrameworkConfig()
            client = APIClient()
            mock_driver = MagicMock()
            page = BasePage(mock_driver)
            factory = DataManager()
            logger = get_logger("performance_test")
            
            end_time = time.time()
            initialization_time = end_time - start_time
            
        with allure.step("Verify performance"):
            # Should initialize quickly (less than 1 second)
            assert initialization_time < 1.0
            
            allure.attach(f"Initialization time: {initialization_time:.3f}s", "Performance", allure.attachment_type.TEXT)
            
        with allure.step("Test data generation performance"):
            start_time = time.time()
            
            # Generate multiple data samples
            for _ in range(10):
                factory.generate_test_dataset(5, 3, 5)
                
            end_time = time.time()
            generation_time = end_time - start_time
            
            # Should generate data quickly
            assert generation_time < 1.0
            
            allure.attach(f"Data generation time (100 samples): {generation_time:.3f}s", "Performance", allure.attachment_type.TEXT)
